backend/requirements.txt
fastapi
uvicorn
python-dotenv
langchain
langchain-google-genai
langchain-openai
langgraph
pydantic
duckduckgo-search
beautifulsoup4
langchain-community
ddgs